Extract ty.rs, expr.rs
