+++ src/hir/expr.rs
@@ -1,0 +1,273 @@
+use crate::hir::{FunTy, Ty};
+
+pub type Typed<T> = (T, Ty);
+pub type TypedExpr = Typed<Expr>;
+
+#[derive(Debug, Clone)]
+pub enum Expr {
+    Number(i64),
+    PseudoVar(PseudoVar),
+    LVarRef(String),
+    ArgRef(usize),
+    FuncRef(String),
+    OpCall(String, Box<Typed<Expr>>, Box<Typed<Expr>>),
+    FunCall(Box<Typed<Expr>>, Vec<Typed<Expr>>),
+    If(Box<Typed<Expr>>, Vec<Typed<Expr>>, Vec<Typed<Expr>>),
+    Yield(Box<Typed<Expr>>),
+    While(Box<Typed<Expr>>, Vec<Typed<Expr>>),
+    Alloc(String),
+    Assign(String, Box<Typed<Expr>>),
+    Return(Box<Typed<Expr>>),
+    Cast(CastType, Box<Typed<Expr>>),
+    // Appears after `lower_async_if`
+    CondReturn(
+        Box<Typed<Expr>>,
+        Box<Typed<Expr>>,
+        Vec<Typed<Expr>>,
+        Box<Typed<Expr>>,
+        Vec<Typed<Expr>>,
+    ),
+    // Appears after `lower_if`
+    Br(Box<Typed<Expr>>, usize),
+    CondBr(Box<Typed<Expr>>, usize, usize),
+    BlockArgRef,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub enum PseudoVar {
+    True,
+    False,
+    Null,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub enum CastType {
+    AnyToFun(FunTy),
+    AnyToInt,
+    NullToAny,
+    IntToAny,
+    FunToAny,
+}
+
+impl CastType {
+    pub fn result_ty(&self) -> Ty {
+        match self {
+            CastType::AnyToFun(x) => x.clone().into(),
+            CastType::AnyToInt => Ty::Int,
+            CastType::NullToAny | CastType::IntToAny | CastType::FunToAny => Ty::Any,
+        }
+    }
+}
+
+impl std::fmt::Display for Expr {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Expr::Number(n) => write!(f, "{}", n),
+            Expr::PseudoVar(PseudoVar::True) => write!(f, "true"),
+            Expr::PseudoVar(PseudoVar::False) => write!(f, "false"),
+            Expr::PseudoVar(PseudoVar::Null) => write!(f, "null"),
+            Expr::LVarRef(name) => write!(f, "{}", name),
+            Expr::ArgRef(idx) => write!(f, "%arg_{}", idx),
+            Expr::FuncRef(name) => write!(f, "{}", name),
+            Expr::OpCall(op, lhs, rhs) => write!(f, "({} {} {})", lhs.0, op, rhs.0),
+            Expr::FunCall(func, args) => {
+                let Ty::Fun(fun_ty) = &func.1 else {
+                    panic!("[BUG] not a function: {:?}", func);
+                };
+                write!(f, "{}{}(", func.0, fun_ty.asyncness)?;
+                for (i, arg) in args.iter().enumerate() {
+                    if i != 0 {
+                        write!(f, ", ")?;
+                    }
+                    write!(f, "{}", arg.0)?;
+                }
+                write!(f, ")")
+            }
+            Expr::If(cond, then, else_) => {
+                write!(f, "if({}){{\n", cond.0)?;
+                for stmt in then {
+                    write!(f, "    {}\n", stmt.0)?;
+                }
+                write!(f, "  }}")?;
+                if !else_.is_empty() {
+                    write!(f, " else {{\n")?;
+                    for stmt in else_ {
+                        write!(f, "    {}\n", stmt.0)?;
+                    }
+                    write!(f, "  }}")?;
+                }
+                Ok(())
+            }
+            Expr::Yield(e) => write!(f, "yield {}  # {}", e.0, e.1),
+            Expr::While(cond, body) => {
+                write!(f, "while {} {{\n", cond.0)?;
+                for stmt in body {
+                    write!(f, "  {}\n", stmt.0)?;
+                }
+                write!(f, "}}")
+            }
+            Expr::Alloc(name) => write!(f, "alloc {}", name),
+            Expr::Assign(name, e) => write!(f, "{} = {}", name, e.0),
+            Expr::Return(e) => write!(f, "return {}  # {}", e.0, e.1),
+            Expr::Cast(cast_type, e) => write!(f, "({} as {})", e.0, cast_type.result_ty()),
+            Expr::CondReturn(cond, fexpr_t, _args_t, fexpr_f, _args_f) => {
+                let Ty::Fun(fun_ty_t) = &fexpr_t.1 else {
+                    panic!("[BUG] not a function: {:?}", fexpr_t);
+                };
+                let Ty::Fun(fun_ty_f) = &fexpr_f.1 else {
+                    panic!("[BUG] not a function: {:?}", fexpr_f);
+                };
+                write!(
+                    f,
+                    "cond_return {}, {}{}(...), {}{}(...)",
+                    cond.0, fexpr_t.0, fun_ty_t.asyncness, fexpr_f.0, fun_ty_f.asyncness
+                )
+            }
+            Expr::Br(e, target) => write!(f, "%br ^bb{}({})  # {}", target, e.0, e.1),
+            Expr::CondBr(cond, target_t, target_f) => {
+                write!(f, "%cond_br {} ^bb{} ^bb{}", cond.0, target_t, target_f)
+            }
+            Expr::BlockArgRef => write!(f, "%block_arg"),
+        }
+    }
+}
+
+impl Expr {
+    pub fn number(n: i64) -> TypedExpr {
+        (Expr::Number(n), Ty::Int)
+    }
+
+    //pub fn pseudo_var(pv: PseudoVar) -> TypedExpr {
+    //    let t = match pv {
+    //        PseudoVar::True | PseudoVar::False => Ty::Bool,
+    //        PseudoVar::Null => Ty::Null,
+    //    };
+    //    (Expr::PseudoVar(pv), t)
+    //}
+
+    pub fn lvar_ref(name: impl Into<String>, ty: Ty) -> TypedExpr {
+        (Expr::LVarRef(name.into()), ty)
+    }
+
+    pub fn arg_ref(idx: usize, ty: Ty) -> TypedExpr {
+        (Expr::ArgRef(idx), ty)
+    }
+
+    pub fn func_ref(name: impl Into<String>, fun_ty: FunTy) -> TypedExpr {
+        (Expr::FuncRef(name.into()), fun_ty.into())
+    }
+
+    pub fn op_call(op_: impl Into<String>, lhs: TypedExpr, rhs: TypedExpr) -> TypedExpr {
+        let op = op_.into();
+        let ty = match &op[..] {
+            "+" | "-" | "*" | "/" => Ty::Int,
+            "<" | "<=" | ">" | ">=" | "==" | "!=" => Ty::Bool,
+            _ => panic!("[BUG] unknown operator: {op}"),
+        };
+        (Expr::OpCall(op, Box::new(lhs), Box::new(rhs)), ty)
+    }
+
+    pub fn fun_call(func: TypedExpr, args: Vec<TypedExpr>) -> TypedExpr {
+        let result_ty = match &func.1 {
+            Ty::Fun(f) => *f.ret_ty.clone(),
+            _ => panic!("[BUG] not a function: {:?}", func),
+        };
+        (Expr::FunCall(Box::new(func), args), result_ty)
+    }
+
+    pub fn if_(cond: TypedExpr, then: Vec<TypedExpr>, else_: Vec<TypedExpr>) -> TypedExpr {
+        if cond.1 != Ty::Bool {
+            panic!("[BUG] if cond not bool: {:?}", cond);
+        }
+        let t1 = yielded_ty(&then);
+        let t2 = yielded_ty(&else_);
+        if t1 != t2 {
+            panic!("[BUG] if types mismatch (t1: {:?}, t2: {:?})", t1, t2);
+        }
+
+        (Expr::If(Box::new(cond), then, else_), t1)
+    }
+
+    pub fn yield_(expr: TypedExpr) -> TypedExpr {
+        let t = expr.1.clone();
+        (Expr::Yield(Box::new(expr)), t)
+    }
+
+    pub fn yield_null() -> TypedExpr {
+        let null = (Expr::PseudoVar(PseudoVar::Null), Ty::Null);
+        (Expr::Yield(Box::new(null)), Ty::Null)
+    }
+
+    pub fn while_(cond: TypedExpr, body: Vec<TypedExpr>) -> TypedExpr {
+        if cond.1 != Ty::Bool {
+            panic!("[BUG] while cond not bool: {:?}", cond);
+        }
+        (Expr::While(Box::new(cond), body), Ty::Null)
+    }
+
+    pub fn assign(name: impl Into<String>, e: TypedExpr) -> TypedExpr {
+        (Expr::Assign(name.into(), Box::new(e)), Ty::Void)
+    }
+
+    pub fn return_(e: TypedExpr) -> TypedExpr {
+        (Expr::Return(Box::new(e)), Ty::Void)
+    }
+
+    pub fn cast(cast_type: CastType, e: TypedExpr) -> TypedExpr {
+        let ty = match &cast_type {
+            CastType::AnyToFun(f) => f.clone().into(),
+            CastType::AnyToInt => Ty::Int,
+            CastType::NullToAny => Ty::Any,
+            CastType::IntToAny => Ty::Any,
+            CastType::FunToAny => Ty::Any,
+        };
+        (Expr::Cast(cast_type, Box::new(e)), ty)
+    }
+
+    pub fn cond_return(
+        cond: TypedExpr,
+        fexpr_t: TypedExpr,
+        args_t: Vec<TypedExpr>,
+        fexpr_f: TypedExpr,
+        args_f: Vec<TypedExpr>,
+    ) -> TypedExpr {
+        (
+            Expr::CondReturn(
+                Box::new(cond),
+                Box::new(fexpr_t),
+                args_t,
+                Box::new(fexpr_f),
+                args_f,
+            ),
+            Ty::Void,
+        )
+    }
+
+    pub fn br(value: TypedExpr, target: usize) -> TypedExpr {
+        (Expr::Br(Box::new(value), target), Ty::Void)
+    }
+
+    pub fn cond_br(cond: TypedExpr, target_t: usize, target_f: usize) -> TypedExpr {
+        (Expr::CondBr(Box::new(cond), target_t, target_f), Ty::Void)
+    }
+
+    pub fn block_arg_ref(ty: Ty) -> TypedExpr {
+        (Expr::BlockArgRef, ty)
+    }
+
+    pub fn is_async_fun_call(&self) -> bool {
+        match self {
+            Expr::FunCall(fexpr, _args) => fexpr.1.is_async_fun(),
+            _ => false,
+        }
+    }
+}
+
+pub fn yielded_ty(stmts: &[TypedExpr]) -> Ty {
+    let stmt = stmts.last().unwrap();
+    match &stmt.0 {
+        Expr::Yield(val) => val.1.clone(),
+        Expr::Return(_) => Ty::Void,
+        _ => panic!("[BUG] if branch not terminated with yield: {:?}", stmt),
+    }
+}
